\name{BrCa443}
\docType{data}
\alias{BrCa443}
\title{Breast Cancer 443 Data Set}

\description{
BrCa443 is an ExpressionSet object that contains gene expression values for 7 genes, RKIP, HMGA2, SPP1, CXCR4, MMP1, metaBACH1, and metaLET7 for 443 breast cancer patients. It also contains paired survival data for each patient in the form of survival and event data.
}

\usage{data(BrCa443)}
\format{ExpressionSet with expression data and survival data}
\source{GSE5327, GSE2034, and GSE2603}
