\name{sigsquared}
\alias{sigsquared}

\title{The sigsquared package}

\description{
The sigsquared package attempts to detect the presence of alternate signaling states of an input pathway that significantly predict differential survival outcome within a mixed cohort of patients. The main goal of this package is to generate gene signatures given known signaling pathways that are predictive of differential survival outcome. The two main functions used to accomplish this goal are first, the ability to train model parameters for a given linear network model, and second, the ability to apply the model and trained parameters to transcript data.
}
